Update and rename bfs.py to BFS.py
